Conflicts:
	api/src/org/labkey/api/data/TableViewForm.java
	api/src/org/labkey/api/webdav/FileSystemResource.java
	specimen/src/org/labkey/specimen/actions/SpecimenController.java

+++ study/src/org/labkey/study/assay/AssayPublishConfirmAction.java
@@ -61,7 +61,7 @@ public class AssayPublishConfirmAction extends AbstractPublishConfirmAction<Assa
             if (_protocol == null)
             {
                 if (getRowId() != null)
-                    _protocol = ExperimentService.get().getExpProtocol(getRowId().intValue());
+                    _protocol = ExperimentService.get().getExpProtocol(getRowId());
                 else
                     throw new NotFoundException("Protocol ID not specified.");
